[ISSUE-1103] Move smart API to api directory
