+++ api/smart/api/generate.go
