+++ api/smart/api/smart.yaml
